GestionProduit_crud_+_controle_saisie

+++ src/Controller/ProduitController.php
@@ -10,22 +10,23 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
-use Psr\Log\LoggerInterface;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 #[Route('/produit')]
 final class ProduitController extends AbstractController
 {
-    private LoggerInterface $logger;
-
-    public function __construct(LoggerInterface $logger)
-    {
-        $this->logger = $logger;
-    }
-
     #[Route(name: 'app_produit_index', methods: ['GET'])]
     public function index(ProduitRepository $produitRepository): Response
     {
         return $this->render('produit/index.html.twig', [
+            'produits' => $produitRepository->findAll(),
+        ]);
+    }
+    
+    #[Route('/dbfrsproduit', name: 'dbfrsproduit', methods: ['GET'])]
+    public function indexfrs(ProduitRepository $produitRepository): Response
+    {
+        return $this->render('dbFournisseurProduits.html.twig', [
             'produits' => $produitRepository->findAll(),
         ]);
     }
@@ -37,59 +38,47 @@ final class ProduitController extends AbstractController
         $form = $this->createForm(ProduitType::class, $produit);
         $form->handleRequest($request);
 
-        if ($form->isSubmitted()) {
-            $this->logger->info('Form submitted', ['data' => $form->getData()]);
-        }
-
         if ($form->isSubmitted() && $form->isValid()) {
-            // Ensure Fournisseur is set
-            /*if ($produit->getFournisseur() === null) {
-                $this->addFlash('error', 'Fournisseur must be set.');
-                return $this->redirectToRoute('app_produit_new');
-            }*/
             $entityManager->persist($produit);
             $entityManager->flush();
 
-            $this->logger->info('Produit saved', ['produit' => $produit]);
-
-            return $this->redirectToRoute('app_produit_index');
+            return $this->redirectToRoute('dbfrsproduit', [], Response::HTTP_SEE_OTHER);
         }
 
         return $this->render('produit/new.html.twig', [
             'produit' => $produit,
-            'form' => $form->createView(),
+            'form' => $form,
         ]);
     }
 
-    #[Route('/{id}', name: 'app_produit_show', methods: ['GET'])]
-    public function show(Produit $produit): Response
+    #[Route('/{id<\d+>}', name: 'app_produit_show', methods: ['GET'])]
+    public function show(int $id, ProduitRepository $produitRepository): Response
     {
+$produit = $produitRepository->find($id);
+        if (!$produit) {
+            throw $this->createNotFoundException('Produit not found');
+        }
+
         return $this->render('produit/show.html.twig', [
             'produit' => $produit,
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'app_produit_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Produit $produit, EntityManagerInterface $entityManager): Response
+    #[Route('/{id<\d+>}/edit', name: 'app_produit_edit', methods: ['GET', 'POST'])]
+    public function edit(Request $request, int $id, ProduitRepository $produitRepository, EntityManagerInterface $entityManager): Response
     {
+$produit = $produitRepository->find($id);
+        if (!$produit) {
+            throw $this->createNotFoundException('Produit not found');
+        }
+
         $form = $this->createForm(ProduitType::class, $produit);
         $form->handleRequest($request);
 
-        if ($form->isSubmitted()) {
-            $this->logger->info('Form submitted', ['data' => $form->getData()]);
-        }
-
         if ($form->isSubmitted() && $form->isValid()) {
-            // Ensure Fournisseur is set
-            if ($produit->getFournisseur() === null) {
-                $this->addFlash('error', 'Fournisseur must be set.');
-                return $this->redirectToRoute('app_produit_edit', ['id' => $produit->getId()]);
-            }
             $entityManager->flush();
 
-            $this->logger->info('Produit updated', ['produit' => $produit]);
-
-            return $this->redirectToRoute('app_produit_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('dbfrsproduit', [], Response::HTTP_SEE_OTHER);
         }
 
         return $this->render('produit/edit.html.twig', [
@@ -98,15 +87,21 @@ final class ProduitController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_produit_delete', methods: ['POST'])]
-    public function delete(Request $request, Produit $produit, EntityManagerInterface $entityManager): Response
+    #[Route('/{id<\d+>}', name: 'app_produit_delete', methods: ['POST'])]
+    public function delete(Request $request, int $id, ProduitRepository $produitRepository, EntityManagerInterface $entityManager): Response
     {
+$produit = $produitRepository->find($id);
+        if (!$produit) {
+            throw $this->createNotFoundException('Produit not found');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$produit->getId(), $request->request->get('_token'))) {
             $entityManager->remove($produit);
             $entityManager->flush();
         }
 
-        return $this->redirectToRoute('app_produit_index', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('dbfrsproduit', [], Response::HTTP_SEE_OTHER);
     }
-}
 
+
+}
